Move save_weights_to_checkpoint to shared utils
