+++ src/maxtext/checkpoint_conversion/standalone_scripts/convert_qwen3_moe.py
@@ -32,7 +32,7 @@ import torch
 from safetensors import safe_open
 from tqdm import tqdm
 
-from maxtext.checkpoint_conversion.standalone_scripts import llama_or_mistral_ckpt
+from maxtext.checkpoint_conversion.utils.utils import save_weights_to_checkpoint
 from maxtext.inference.inference_utils import str2bool
 from maxtext.utils import max_logging
 
@@ -271,7 +271,7 @@ def main(args):
   max_logging.log(f"Starting conversion for Qwen3-MoE model size: {args.model_size}")
   jax_weights = convert_hf_to_maxtext(args.base_model_path, model_params)
   max_logging.log(f"Conversion complete. Saving MaxText checkpoint to {args.maxtext_model_path}")
-  llama_or_mistral_ckpt.save_weights_to_checkpoint(
+  save_weights_to_checkpoint(
       args.maxtext_model_path, jax_weights, args.simulated_cpu_devices_count, args.use_ocdbt, args.use_zarr3
   )
   max_logging.log("Checkpoint saved successfully.")
