+++ src/maxtext/checkpoint_conversion/standalone_scripts/llama4_ckpt_unscanned.py
@@ -55,7 +55,8 @@ from safetensors import safe_open
 import torch
 from tqdm import tqdm
 
-from maxtext.checkpoint_conversion.standalone_scripts.llama_or_mistral_ckpt import save_weights_to_checkpoint, MODEL_PARAMS_DICT
+from maxtext.checkpoint_conversion.standalone_scripts.llama_or_mistral_ckpt import MODEL_PARAMS_DICT
+from maxtext.checkpoint_conversion.utils.utils import save_weights_to_checkpoint
 from maxtext.inference.inference_utils import str2bool
 from maxtext.utils import max_logging
